# docs/api/openapi_v1.yaml
openapi: 3.0.3
info:
  title: CHIQUISTRUKIS API
  version: 1.0.0
  description: >
    Control de asistencia biométrica. Auth Bearer (Redis).
    Paths in English /api/v1 (SB-04). Domain rules from Tema 5.
servers:
  - url: http://127.0.0.1:8000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /api/v1/biometric-imports:
    get:
      operationId: listBiometricImports
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [draft, confirmed, cancelled] }
        - in: query
          name: month
          schema: { type: integer }
        - in: query
          name: year
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200': { description: Import history }
    post:
      operationId: createBiometricImport
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Draft import with rows (matched/new)
  /api/v1/biometric-imports/{id}:
    get:
      operationId: getBiometricImport
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Import + rows }
  /api/v1/biometric-imports/{id}/rows/{rowId}:
    patch:
      operationId: updateImportRow
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: rowId
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dni: { type: string }
                last_names: { type: string }
                first_names: { type: string }
                action:
                  type: string
                  enum: [research, register_new, skip]
      responses:
        '200': { description: Row updated }
        '409': { description: Not draft }
  /api/v1/biometric-imports/{id}/confirmation:
    post:
      operationId: confirmBiometricImport
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Confirmed; consolidates attendance_day }
        '400': { description: Unresolved new rows }
        '409': { description: Not draft }
  /api/v1/biometric-imports/{id}/cancellation:
    post:
      operationId: cancelBiometricImport
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200': { description: Cancelled; reverts marks and recalculates }
        '409': { description: Not confirmed }
...